Reorganize ACExceptionFinder implementation
ACExceptionFinder class shouldn't be accessible by user, so it was
moved into algos::algebraic_constraints namespace. ACException was
separated from ACExceptionFinder. ValueBelongsToRanges was moved
to ACExceptionFinder
